some comments and now used long instead of int

+++ src/ethModul01/Doppeln_Eingabe.java
@@ -9,12 +9,18 @@ public class Doppeln_Eingabe {
     public static void main(String[] args) {
         Scanner eingabe = new Scanner(System.in);
 
-        //Deklaration der Variablen als int
-        int zahl;
-        int doppelte_zahl;
+        //Initialisierung der Variablen als long
+        long zahl;
+        long doppelte_zahl;
+
+        //Eingabe durch Benutzer des zu errechnenden Werts
         System.out.println("Bitte geben Sie den Wert ein, welcher verdoppelt werden soll:");
-        zahl = eingabe.nextInt();
+
+        //Berechnung des doppelten Werts
+        zahl = eingabe.nextLong();
         doppelte_zahl = zahl * 2;
+
+        //Ausgabe des Werts
         System.out.println("Die Eingabe lautete: " + zahl);
         System.out.println("Der errechnete Wert lautet: " + doppelte_zahl);
     }
